% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_schema}
\alias{data_schema}
\title{Input table specification}
\usage{
data_schema()
}
\value{
A named list of table specifications.
}
\description{
Required columns and primary keys for supplied tables. Optional columns and
analysis-specific requirements are described in the schema vignette.
}
