% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_deliberation}
\alias{read_deliberation}
\title{Read named CSV input tables}
\usage{
read_deliberation(path)
}
\arguments{
\item{path}{Directory containing CSV files named for \code{\link[=data_schema]{data_schema()}} tables.}
}
\value{
A \code{deliberation_data} object with file hashes.
}
\description{
Read named CSV input tables
}
